# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SGT-HUAP — API de Gestión de Turnos",
    "description": "API REST del Sistema de Gestión de Turnos del Hospital de Urgencia Asistencia Pública (HUAP). Gestiona funcionarios, servicios, puestos, turnos, planificaciones, rotativas, solicitudes, ofertas, notificaciones, bitácora y estadísticas.\n\nLa autenticación es mediante JWT (Bearer token). Obtén el token en `POST /api/v2/funcionarios/login` y pulsa **Authorize** para probar los endpoints protegidos.",
    "contact": {
      "name": "Equipo SGT-HUAP"
    },
    "license": {
      "name": "Uso interno HUAP"
    },
    "version": "v2"
  },
  "servers": [
    {
      "url": "/",
      "description": "Servidor actual"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v2/reglas-servicio/{id}": {
      "put": {
        "tags": [
          "regla-servicio-controller"
        ],
        "operationId": "actualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReglaServicioDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "regla-servicio-controller"
        ],
        "operationId": "eliminar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v2/planificaciones/{id}": {
      "get": {
        "tags": [
          "planificacion-controller"
        ],
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanificacionDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "planificacion-controller"
        ],
        "operationId": "actualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanificacionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanificacionDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "planificacion-controller"
        ],
        "operationId": "eliminar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/ofertas-generales/{id}/seleccionar/{idPostulacion}": {
      "put": {
        "tags": [
          "oferta-general-controller"
        ],
        "operationId": "seleccionar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idPostulacion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idJefatura",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaGeneralEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/funcionarios/{id}": {
      "put": {
        "tags": [
          "Funcionarios y autenticación"
        ],
        "summary": "Actualizar un funcionario",
        "description": "Un usuario solo puede modificar su propio registro; solo JEFATURA/ADMINISTRADOR pueden cambiar el rol o el estado de otro funcionario.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Funcionario actualizado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para modificar a otro funcionario o su rol/estado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Funcionario no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/planificaciones/{id}/generar": {
      "post": {
        "tags": [
          "planificacion-controller"
        ],
        "operationId": "generar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/funcionarios/switch-service": {
      "post": {
        "tags": [
          "Funcionarios y autenticación"
        ],
        "summary": "Cambiar de servicio en la sesión activa",
        "description": "Genera un nuevo JWT para otro servicio al que el funcionario autenticado tenga acceso.",
        "operationId": "switchService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/funcionarios/register/{rut}": {
      "post": {
        "tags": [
          "Funcionarios y autenticación"
        ],
        "summary": "Registrar personal como funcionario",
        "description": "Registra en el sistema de turnos a una persona existente en el personal (viewPersonal). Requiere rol JEFATURA o ADMINISTRADOR.",
        "operationId": "registerPersonal",
        "parameters": [
          {
            "name": "rut",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Funcionario registrado; retorna el nuevo ID",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "400": {
            "description": "RUT inválido o datos faltantes",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "409": {
            "description": "El funcionario ya estaba registrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/funcionarios/login": {
      "post": {
        "tags": [
          "Funcionarios y autenticación"
        ],
        "summary": "Autenticación (paso 1)",
        "description": "Valida RUT y contraseña. Si el usuario tiene varios servicios, devuelve un token de pre-autorización y la lista de servicios para elegir en el paso 2. Incluye bloqueo por intentos fallidos.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credenciales válidas; retorna token de pre-autorización y servicios disponibles",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Credenciales incompletas",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Usuario sin servicios asignados",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "description": "Cuenta bloqueada por demasiados intentos fallidos",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v2/funcionarios/login/select-service": {
      "post": {
        "tags": [
          "Funcionarios y autenticación"
        ],
        "summary": "Autenticación (paso 2): seleccionar servicio",
        "description": "Recibe el token de pre-autorización y el servicio elegido, y devuelve el JWT definitivo (con rol de sistema y rol de servicio) junto con el perfil.",
        "operationId": "selectService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada; retorna JWT y perfil",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Token de pre-autorización inválido o expirado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v2/turnos/funcionario/{idFuncionario}": {
      "get": {
        "tags": [
          "turno-controller"
        ],
        "operationId": "getTurnosByFuncionario",
        "parameters": [
          {
            "name": "idFuncionario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReglaServicioDTO": {
        "type": "object",
        "properties": {
          "idRegla": {
            "type": "integer",
            "format": "int64"
          },
          "idServicio": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "aplicaFinDeSemana": {
            "type": "boolean"
          },
          "aplicaFeriado": {
            "type": "boolean"
          },
          "tiempoMinutos": {
            "type": "integer",
            "format": "int32"
          },
          "idTipoTurnoInicio": {
            "type": "integer",
            "format": "int64"
          },
          "idTipoTurnoFin": {
            "type": "integer",
            "format": "int64"
          },
          "nombreTipoTurnoInicio": {
            "type": "string"
          },
          "nombreTipoTurnoFin": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Introduce el token JWT obtenido en el login (sin el prefijo 'Bearer ').",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}